+++ TimberCommons/package/CHANGELOG.md
@@ -5,6 +5,7 @@
 * Add a feature control file (TimberDev_Features.txt).
 * Add a feature to show duration as "Xd Yh" (instade of simple "XX hours") in the good consuming buidlings.
 * Add a feature to show duration as "Xd Yh" (instead of rounding to teh closes whoel day) in the growable dialogs.
+* Add `PrefabOptimizer.MaxExpectedRegistrySize` feature to override the value and stop logs spam. Disabled by default.
 
 # v1.4 (27 Jan 2024)
 * [Fix #28] "Water bump" at the water intake.